Start moving admin/username and admin/password into their own controllers
