+++ app/routes/profile/admin.js
@@ -8,6 +8,6 @@ import SessionMixin from 'gistr/mixins/session';
 export default Ember.Route.extend(RestrictedRouteMixin, FormRouteMixin, SessionMixin, {
   setupController: function(controller/*, model*/) {
     var user = this.get('currentUser');
-    controller.setProperties(user.getProperties('username'));
+    controller.set('controllers.profile/admin/username.username', user.get('username'));
   }
 });
