+++ app/templates/profile/admin/password.hbs
@@ -1,0 +1,1 @@
+<h2>Change password</h2>
